Remove theme.js and declare theme in _app.js
