+++ pages/_app.js
@@ -1,5 +1,18 @@
-import { ChakraProvider } from '@chakra-ui/react';
-import theme from './theme';
+import { ChakraProvider, extendTheme } from '@chakra-ui/react';
+
+const theme = extendTheme({
+  styles: {
+    global: () => ({
+      body: {
+        bg: '#171313',
+      },
+    }),
+  },
+  fonts: {
+    heading: 'Josefin Sans, sans-serif',
+    body: 'Josefin Sans, sans-serif',
+  },
+});
 
 function MyApp({ Component, pageProps }) {
   return (
